+++ packages/uniwind/tests/consts.ts
@@ -8,10 +8,13 @@ export const TW_SPACING = 4
 
 export const SAFE_AREA_INSET_TOP = 21
 export const SAFE_AREA_INSET_BOTTOM = 42
+export const SAFE_AREA_INSET_LEFT = 7
+export const SAFE_AREA_INSET_RIGHT = 13
 
 export const SCREEN_WIDTH = 390
 export const SCREEN_HEIGHT = 844
 
 export const UNIWIND_CONTEXT_MOCK = {
     scopedTheme: null,
+    rtl: null,
 } satisfies UniwindContextType